moved tests to modules
